Rename the DatabaseModels folder to Entities and move it from Modles to Data. Then rename the Models.DatabaseModels namespace in all project files to Data.Entities.
