+++ src/Open-School-Library/Data/Entities/Book.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Open_School_Library.Models.DatabaseModels
+namespace Open_School_Library.Data.Entities
 {
     public class Book
     {
