+++ src/Open-School-Library/Data/Entities/Genre.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Open_School_Library.Models.DatabaseModels
+namespace Open_School_Library.Data.Entities
 {
     public class Genre
     {
