+++ src/Open-School-Library/Data/Entities/Student.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Open_School_Library.Models.DatabaseModels
+namespace Open_School_Library.Data.Entities
 {
     public class Student
     {
