+++ src/Open-School-Library/Models/BookViewModels/BookEditViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc.Rendering;
-using Open_School_Library.Models.DatabaseModels;
+using Open_School_Library.Data.Entities;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
