+++ src/Open-School-Library/Models/StudentViewModels/StudentIndexViewModel.cs
@@ -1,4 +1,4 @@
-﻿using Open_School_Library.Models.DatabaseModels;
+﻿using Open_School_Library.Data.Entities;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
