+++ src/Open-School-Library/Views/Books/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<Open_School_Library.Models.DatabaseModels.Book>
+@model IEnumerable<Open_School_Library.Data.Entities.Book>
 
 <h2>Book Catalog</h2>
