+++ src/Open-School-Library/Views/Settings/Edit.cshtml
@@ -1,4 +1,4 @@
-@model Open_School_Library.Models.DatabaseModels.Setting
+@model Open_School_Library.Data.Entities.Setting
 
 @{
     ViewData["Title"] = "Edit";
