+++ src/Open-School-Library/Views/Settings/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<Open_School_Library.Models.DatabaseModels.Setting>
+@model IEnumerable<Open_School_Library.Data.Entities.Setting>
 
 @{
     ViewData["Title"] = "Index";
